+++ Packages/com.nickmaltbie.statemachineunity.netcode/Scripts/NetworkSMAnim.cs
@@ -17,9 +17,10 @@
 // SOFTWARE.
 
 using System;
-using System.Threading;
+using System.Linq;
 using nickmaltbie.StateMachineUnity.Attributes;
 using nickmaltbie.StateMachineUnity.Event;
+using nickmaltbie.StateMachineUnity.Utils;
 using Unity.Netcode;
 using UnityEngine;
 
@@ -57,11 +58,6 @@ namespace nickmaltbie.StateMachineUnity.netcode
                 serializer.SerializeValue(ref lockAnimationTime);
             }
         }
-
-        /// <summary>
-        /// Has the completed event been raised for the current state yet.
-        /// </summary>
-        private int raisedCompletedEvent;
 
         /// <summary>
         /// Pending request for the animator.
@@ -106,6 +102,12 @@ namespace nickmaltbie.StateMachineUnity.netcode
 
             AttachedAnimator ??= gameObject.GetComponent<Animator>();
 
+            if (AttachedAnimator != null)
+            {
+                AnimationCompleteListener listener = AttachedAnimator.gameObject.AddComponent<AnimationCompleteListener>();
+                listener.OnAnimationCompleted += OnAnimationComplete;
+            }
+
             if (IsOwner)
             {
                 UpdateAnimationState();
@@ -139,7 +141,6 @@ namespace nickmaltbie.StateMachineUnity.netcode
                 return;
             }
 
-            raisedCompletedEvent = 0;
             _currentAnimationState.Value = new AnimSMRequestNetwork(req);
 
             if (req.lockAnimationTime > 0)
@@ -185,6 +186,19 @@ namespace nickmaltbie.StateMachineUnity.netcode
             base.Update();
         }
 
+        public void OnAnimationComplete(object source, string clipName)
+        {
+            if (Attribute.GetCustomAttribute(CurrentState, typeof(AnimationAttribute)) is AnimationAttribute animAttr)
+            {
+                AnimatorStateInfo currentState = AttachedAnimator.GetCurrentAnimatorStateInfo(0);
+                AnimatorClipInfo[] animClips = AttachedAnimator.GetCurrentAnimatorClipInfo(0);
+                if (currentState.IsName(animAttr.StateName) && animClips.Any(animClip => animClip.clip.name == clipName))
+                {
+                    RaiseEvent(AnimationCompleteEvent.Instance);
+                }
+            }
+        }
+
         /// <summary>
         /// Update the animation state of this fixed sm animator
         /// based on the current animation tag of the current state of
@@ -194,13 +208,7 @@ namespace nickmaltbie.StateMachineUnity.netcode
         {
             if (Attribute.GetCustomAttribute(CurrentState, typeof(AnimationAttribute)) is AnimationAttribute animAttr)
             {
-                if (AttachedAnimator.GetCurrentAnimatorStateInfo(0).shortNameHash == animAttr.AnimationHash &&
-                    AttachedAnimator.GetCurrentAnimatorStateInfo(0).normalizedTime >= 1 &&
-                    Interlocked.CompareExchange(ref raisedCompletedEvent, 1, 0) == 0)
-                {
-                    RaiseEvent(AnimationCompleteEvent.Instance);
-                }
-                else if (lockUntilTime >= unityService.time)
+                if (lockUntilTime >= unityService.time)
                 {
                     // We are locked, do not cross fade into new animation.
                 }