+++ Packages/com.nickmaltbie.statemachineunity/Tests/EditMode/Fixed/StateMachineAnimTests.cs
@@ -16,6 +16,7 @@
 // ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+using System.Linq;
 using Moq;
 using nickmaltbie.StateMachineUnity.Attributes;
 using nickmaltbie.StateMachineUnity.Fixed;
@@ -198,6 +199,7 @@ namespace nickmaltbie.StateMachineUnity.Tests.EditMode.Fixed
             anim.Play(AnimA, 0, 1.0f);
             anim.Update(1.0f);
             sm.Update();
+            sm.OnAnimationComplete(null, anim.GetCurrentAnimatorClipInfo(0).First().clip.name);
             Assert.AreEqual(typeof(TimeoutState), sm.CurrentState);
         }