+++ Packages/com.nickmaltbie.statemachineunity.netcode/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## In Progress
 
+## [1.2.3] - 2023-02-22
+
+* Patched how OnAnimationComplete events are handled by using events within
+    animation clips.
+
 ## [1.2.2] - 2023-02-18
 
 * Added edge case coverage for initializing state machines owned